Morgan State AI chatbot/BackEnd/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# AI/ML
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
tiktoken==0.5.2

# Vector Store
pinecone-client==3.0.2

# Environment & Config
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1